feat: move filter stories to filter/stories
